Student edit Page

+++ src/main/java/com/albany/edu/fwp/action/StudentAction.java
@@ -25,6 +25,7 @@ import com.albany.edu.fwp.dao.MealCourseDAO;
 import com.albany.edu.fwp.dao.QuadInfoDAO;
 import com.albany.edu.fwp.dao.StudentDAO;
 import com.albany.edu.fwp.model.FoodItems;
+import com.albany.edu.fwp.model.FoodSelected;
 import com.albany.edu.fwp.model.MealCourse;
 import com.albany.edu.fwp.model.QuadInfo;
 import com.albany.edu.fwp.model.Student;
@@ -46,7 +47,9 @@ public class StudentAction extends ActionSupport {
     private List<String> mealCourseNames;
     private Student student;
     private FoodItems foodItems;
+    private MealCourse mealCourse;
     private HashMap<String, HashMap<String, List<List<String>>>> allFoodItems;
+    private HashMap<String, List<List<String>>> selectedFoodItemsPerMealCourseMap;
  
     public void setFoodItemsDAO(FoodItemsDAO foodItemsDAO) {
 		this.foodItemsDAO = foodItemsDAO;
@@ -72,10 +75,15 @@ public class StudentAction extends ActionSupport {
 //		}
     	//ActionContext.getContext().getParameters();
     	
+    	
+    	String returnString="";
     	session = ServletActionContext.getRequest().getSession();
     	session.setAttribute("studentId", "as132736");
-    	
-		listQuad = quadInfoDAO.list();
+    	request = ServletActionContext.getRequest();
+    	String studentId=session.getAttribute("studentId").toString(); 
+    	student = studentDAO.getStudent(studentId);
+    	List<FoodSelected> listFoodSelected = foodSelectedDAO.listFoodSelected(student);
+    	listQuad = quadInfoDAO.list();
 		quadNames = new ArrayList<String>();
 			for (QuadInfo quad : listQuad){
 				quadNames.add(quad.getQuadName());
@@ -87,20 +95,44 @@ public class StudentAction extends ActionSupport {
 			for (MealCourse mealCourse : listMealCourse){
 				mealCourseNames.add(mealCourse.getMealCourseName());
 			}
-		
-    	allFoodItems = foodItemsDAO.getAllFoodItemsMap();     
-    	for (List foodItem : allFoodItems.get("Indian").get("Breakfast")){
-		    System.out.println("Quad Ids: "+ foodItem);
-		}
-    	
-    	request = ServletActionContext.getRequest();
-    	if( !(request.getParameterMap().isEmpty()) ) { 
-    		String studentId=session.getAttribute("studentId").toString();
-    		student = studentDAO.getStudent(studentId);
-    		if( !(foodSelectedDAO.listFoodSelected(student).isEmpty()) ){
-    			foodSelectedDAO.deleteStudentSelection(student);
-    		}
+			selectedFoodItemsPerMealCourseMap = new HashMap<String, List<List<String>>>();
+			HashMap<String, String> selectedFoodItems = new HashMap<String, String>(); // Only purpose is to pre select radio buttons id student wants to edit his choices
+			HashMap<String, String> selectedFoodItemsQuadAndMealCourse = new HashMap<String, String>(); // Only purpose is to pre select radio buttons id student wants to edit his choices
+			QuadInfo quadInfo;
+    	if( !listFoodSelected.isEmpty()	&& request.getParameterMap().isEmpty() )    	  
+    	{
+    		System.out.println("Edit Page---------->");    		
     		
+    		for (FoodSelected foodSelected : listFoodSelected){ 
+    			List<String> foodItemsAndQuantity = new ArrayList<String>();   			
+    			foodItems = foodItemsDAO.getFoodItem( Integer.toString(foodSelected.getFoodItems().getFoodItemId()));
+    			foodItemsAndQuantity.add(Integer.toString(foodItems.getFoodItemId()));
+    			foodItemsAndQuantity.add(foodItems.getFoodItemName());
+    			foodItemsAndQuantity.add(Integer.toString(foodSelected.getNumberOfPlates()));
+    			foodItemsAndQuantity.add(Integer.toString(foodItems.getCalories()));
+    			selectedFoodItems.put(Integer.toString(foodItems.getFoodItemId()), Integer.toString(foodSelected.getNumberOfPlates()));
+    			mealCourse = mealCourseDAO.getMealCourse(foodItems.getMealCourse().getMealCourseId()); 
+    			quadInfo = quadInfoDAO.getQuadInfo(foodItems.getQuadInfo().getQuadId());
+    			selectedFoodItemsQuadAndMealCourse.put(quadInfo.getQuadName()+mealCourse.getMealCourseName(), mealCourse.getMealCourseName());
+    			if(selectedFoodItemsPerMealCourseMap.get(mealCourse.getMealCourseName())==null){
+    				List<List<String>> foodItemsInMealCourse = new ArrayList<List<String>>();
+    				foodItemsInMealCourse.add(foodItemsAndQuantity);
+    				selectedFoodItemsPerMealCourseMap.put(mealCourse.getMealCourseName(), foodItemsInMealCourse);
+    			}else{
+    				selectedFoodItemsPerMealCourseMap.get(mealCourse.getMealCourseName()).add(foodItemsAndQuantity);
+    			}
+    			
+    			
+    		}
+    		session.setAttribute("selectedFoodItemsIDAndQuantity", selectedFoodItems);
+    		session.setAttribute("selectedFoodItemsQuadAndMealCourse", selectedFoodItemsQuadAndMealCourse);
+    		returnString="edit";
+		}
+    	else if( !(request.getParameterMap().isEmpty()) && request.getParameter("next")==null )
+    	{			   		
+    		if( !(listFoodSelected.isEmpty()) ){
+    			foodSelectedDAO.deleteStudentSelection(student);
+    		}    		
     		DateFormat dateFormat = new SimpleDateFormat("HH:mm:ss");
     		Date date = new Date();
     		dateFormat.format(date).toString();
@@ -112,9 +144,42 @@ public class StudentAction extends ActionSupport {
     			foodItems=foodItemsDAO.getFoodItem(entry.getKey());
     			foodSelectedDAO.insertFoodSelected(Integer.parseInt(entry.getValue()[0]), foodItems, student, dateFormat.format(date).toString());
     		}
-    		//System.out.println("------Parameters----->"+request.getParameterMap().get("5")[0]);
+    		listFoodSelected = foodSelectedDAO.listFoodSelected(student);
+    		for (FoodSelected foodSelected : listFoodSelected){ 
+    			List<String> foodItemsAndQuantity = new ArrayList<String>();   			
+    			foodItems = foodItemsDAO.getFoodItem( Integer.toString(foodSelected.getFoodItems().getFoodItemId()));
+    			foodItemsAndQuantity.add(Integer.toString(foodItems.getFoodItemId()));
+    			foodItemsAndQuantity.add(foodItems.getFoodItemName());
+    			foodItemsAndQuantity.add(Integer.toString(foodSelected.getNumberOfPlates()));  
+    			foodItemsAndQuantity.add(Integer.toString(foodItems.getCalories()));
+    			selectedFoodItems.put(Integer.toString(foodItems.getFoodItemId()), Integer.toString(foodSelected.getNumberOfPlates()));
+    			mealCourse = mealCourseDAO.getMealCourse(foodItems.getMealCourse().getMealCourseId()); 
+    			quadInfo = quadInfoDAO.getQuadInfo(foodItems.getQuadInfo().getQuadId());
+    			selectedFoodItemsQuadAndMealCourse.put(quadInfo.getQuadName()+mealCourse.getMealCourseName(), mealCourse.getMealCourseName());
+    			if(selectedFoodItemsPerMealCourseMap.get(mealCourse.getMealCourseName())==null){
+    				List<List<String>> foodItemsInMealCourse = new ArrayList<List<String>>();
+    				foodItemsInMealCourse.add(foodItemsAndQuantity);
+    				selectedFoodItemsPerMealCourseMap.put(mealCourse.getMealCourseName(), foodItemsInMealCourse);
+    			}else{
+    				selectedFoodItemsPerMealCourseMap.get(mealCourse.getMealCourseName()).add(foodItemsAndQuantity);
+    			}
+    			
+    			//System.out.println("------- "+selectedFoodItemsPerMealCourseMap.get("Breakfast").get(1));
+    		}
+    		session.setAttribute("selectedFoodItemsIDAndQuantity", selectedFoodItems);
+    		session.setAttribute("selectedFoodItemsQuadAndMealCourse", selectedFoodItemsQuadAndMealCourse);
+    		returnString="edit";
+		}
+    	
+    	else{			
+	    	allFoodItems = foodItemsDAO.getAllFoodItemsMap();     
+	    	for (List foodItem : allFoodItems.get("Indian").get("Breakfast")){
+			    System.out.println("Quad Ids: "+ foodItem);
+			}
+	    	returnString="firstchoice";
     	}
-        return SUCCESS;
+    	
+        return returnString;
     }
 
     
@@ -133,5 +198,11 @@ public class StudentAction extends ActionSupport {
     public HashMap<String, HashMap<String, List<List<String>>>> getAllFoodItems() {
           return allFoodItems;
     }
+    
+    public HashMap<String, List<List<String>>> getSelectedFoodItemsPerMealCourseMap() {
+        return selectedFoodItemsPerMealCourseMap;
+    }
+    
+    
 
 }
